Fix: Convert output dimensions to integers for OpenCV VideoWriter
This commit fixes the OpenCV VideoWriter error caused by float dimensions.

Problem:
- Changed scale_factor from int to float for flexible scaling
- output_width/height calculations now produce float values
- OpenCV VideoWriter expects integer dimensions
- Error: "Can't parse 'frameSize'. Sequence item with index 0 has a wrong type"

Solution:
- Explicitly convert output dimensions to int
- output_width = int(metadata['width'] * self.scale_factor)
- output_height = int(metadata['height'] * self.scale_factor)

Changes:
- src/processors/spatial_upscaler.py (lines 133-134)
  - Add int() conversion for output_width
  - Add int() conversion for output_height
  - Add comment explaining OpenCV requirement

Testing:
- test_opencv_dimensions.py
  - Tests various scale factors (2.0x, 1.5x, 2.5x, 0.5x, 4.0x)
  - Confirms all dimensions are integers
  - All 5/5 tests pass ✅

Results:
- Integer dimensions for 2.0x scale: 1920x1080 → 3840x2160 ✅
- Integer dimensions for 1.5x scale: 1920x1080 → 2880x1620 ✅
- Integer dimensions for 4.0x scale: 824x1464 → 3296x5856 ✅
- Integer dimensions for 2.5x scale: 1280x720 → 3200x1800 ✅
- Integer dimensions for 0.5x scale: 3840x2160 → 1920x1080 ✅

Behavior Changes:
- VideoWriter now receives proper integer dimensions
- Fractional scales work correctly
- No more OpenCV type errors

Tested With:
- Windows 11
- Python 3.11.9
- OpenCV 4.11.0
- All scale factors validated

+++ src/processors/spatial_upscaler.py
@@ -129,9 +129,9 @@ class SpatialUpscaler:
                 logger.info(f"Input resolution: {metadata['width']}x{metadata['height']}")
                 logger.info(f"FPS: {metadata['fps']}, Frames: {metadata['frame_count']}")
 
-                # Calculate output resolution
-                output_width = metadata['width'] * self.scale_factor
-                output_height = metadata['height'] * self.scale_factor
+                # Calculate output resolution (convert to int for OpenCV)
+                output_width = int(metadata['width'] * self.scale_factor)
+                output_height = int(metadata['height'] * self.scale_factor)
 
                 logger.info(f"Output resolution: {output_width}x{output_height}")
